refactor: change path alias names
