+++ tsconfig.app.json
@@ -7,16 +7,15 @@
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "module": "ESNext",
     "skipLibCheck": true,
-    "baseUrl": "./src",
+    "baseUrl": ".",
     "paths": {
-      "app/*": ["./app/*"],
-      "processes/*": ["./processes/*"],
-      "pages/*": ["./pages/*"],
-      "widgets/*": ["./widgets/*"],
-      "features/*": ["./features/*"],
-      "entities/*": ["./entities/*"],
-      "shared/*": ["./shared/*"],
-      "src/*": ["./*"]
+      "@app/*": ["./src/app/*"],
+      "@pages/*": ["./src/pages/*"],
+      "@widgets/*": ["./src/widgets/*"],
+      "@features/*": ["./src/features/*"],
+      "@entities/*": ["./src/entities/*"],
+      "@shared/*": ["./src/shared/*"],
+      "@src/*": ["./src/*"]
     },
 
     /* Bundler mode */
